BLD: rename `_directmodule` extension to `_direct`
Fixes odd issue on 32-bit Linux CI job showing:
```
ImportError: dynamic module does not define module export function (PyInit__direct)
```
